Create README.md
This project involves conducting an Explanatory Data Analysis (EDA) on a given dataset of Diwali sales. The steps include importing the dataset, determining its dimensions, printing the first 5 rows, and obtaining basic information about the dataset. Irrelevant or blank columns are dropped, followed by checking for and dropping null values. The data type of the "Amount" column is changed from float to int. Column names are printed and one is renamed. Statistical analysis is performed on the "Age," "Order," and "Amount" columns. Bar charts are used to represent the count of male and female customers. Total purchase amounts are calculated based on gender and age group. Gender counts are obtained. Chart analyses are performed on orders and amounts by state, marital status, occupation, and product category. Finally, a chart analysis is done to identify the top-selling products. The aim is to gain insights into the Diwali sales data, customer demographics, trends, and relationships between variables.